+++ jax/_src/source_info_util.py
@@ -39,8 +39,10 @@ def user_frames(source_info: Optional[Traceback]) -> Iterator[Frame]:
   # We don't use traceback_util.path_starts_with because that incurs filesystem
   # access, which may be slow; we call this function when e.g. adding source
   # provenance annotations to XLA lowerings, so we don't want to incur the cost.
+  # We consider files that end with _test.py as user frames, to allow testing
+  # this mechanism from tests.
   return (x for x in (source_info.frames if source_info else [])
-          if not any(x.file_name.startswith(p) for p in _exclude_paths))
+          if x.file_name.endswith("_test.py") or not any(x.file_name.startswith(p) for p in _exclude_paths))
 
 def user_frame(source_info: Optional[Traceback]) -> Optional[Frame]:
   return next(user_frames(source_info), None)