+++ jax/experimental/jax2tf/tests/jax2tf_test.py
@@ -28,6 +28,8 @@ from jax import test_util as jtu
 from jax.config import config
 from jax.experimental import jax2tf
 from jax.experimental.jax2tf.tests import tf_test_util
+from jax._src import source_info_util
+
 import numpy as np
 import tensorflow as tf  # type: ignore[import]
 
@@ -80,7 +82,7 @@ class Jax2TfTest(tf_test_util.JaxToTfTestCase):
     # Take and return pytrees
     def f_jax(x: Tuple[float, Dict[str, float]]) -> Tuple[float, Dict[str, float]]:
       x_a, x_dict = x
-      return x_a * 2., {k : v * 3. for k, v in x_dict.items()}
+      return x_a * 2., {k: v * 3. for k, v in x_dict.items()}
 
     x = (.7, {"a": .8, "b": .9})
     self.ConvertAndCompare(f_jax, x)
@@ -152,7 +154,6 @@ class Jax2TfTest(tf_test_util.JaxToTfTestCase):
     self.assertAllClose(jnp.sin(big_const),
                         f_conv(tf.constant(big_const)))
 
-
   def test_64bit_behavior_enable_x64(self):
     if not config.jax_enable_x64:
       self.skipTest("requires x64 mode")
@@ -184,9 +185,9 @@ class Jax2TfTest(tf_test_util.JaxToTfTestCase):
     self.ConvertAndCompare(f_jax, 0.7)
 
   @parameterized.named_parameters(jtu.cases_from_list(
-    dict(testcase_name=f"function={with_function}",
-         with_function=with_function)
-    for with_function in [False, True]))
+      dict(testcase_name=f"function={with_function}",
+           with_function=with_function)
+      for with_function in [False, True]))
   def test_gradients_disabled(self, with_function=False):
     f_tf = jax2tf.convert(jnp.tan, with_gradient=False)
     if with_function:
@@ -247,9 +248,9 @@ class Jax2TfTest(tf_test_util.JaxToTfTestCase):
     self.assertAllClose(4., tape.gradient(uv["two"], y))
 
   @parameterized.named_parameters(jtu.cases_from_list(
-    dict(testcase_name=f"function={with_function}",
-         with_function=with_function)
-    for with_function in [False, True]))
+      dict(testcase_name=f"function={with_function}",
+           with_function=with_function)
+      for with_function in [False, True]))
   def test_gradients_with_custom_jvp(self, with_function=True):
     """Check gradients, for a function with custom JVP."""
     @jax.custom_jvp
@@ -281,9 +282,9 @@ class Jax2TfTest(tf_test_util.JaxToTfTestCase):
     self.assertAllClose(3. * 4., tape.gradient(y, x))
 
   @parameterized.named_parameters(jtu.cases_from_list(
-    dict(testcase_name=f"function={with_function}",
-         with_function=with_function)
-    for with_function in [False, True]))
+      dict(testcase_name=f"function={with_function}",
+           with_function=with_function)
+      for with_function in [False, True]))
   def test_gradients_with_custom_vjp(self, with_function=True):
     """Check gradients, for a function with custom VJP."""
     @jax.custom_vjp
@@ -378,6 +379,7 @@ class Jax2TfTest(tf_test_util.JaxToTfTestCase):
 
   def test_custom_jvp(self):
     """Conversion of function with custom JVP"""
+
     @jax.custom_jvp
     def f(x):
       return x * x
@@ -400,6 +402,7 @@ class Jax2TfTest(tf_test_util.JaxToTfTestCase):
 
   def test_custom_vjp(self):
     """Conversion of function with custom VJP"""
+
     @jax.custom_vjp
     def f(x):
       return x * x
@@ -407,6 +410,7 @@ class Jax2TfTest(tf_test_util.JaxToTfTestCase):
     # f_fwd: a -> (b, residual)
     def f_fwd(x):
       return f(x), 3. * x
+
     # f_bwd: (residual, CT b) -> [CT a]
     def f_bwd(residual, ct_b):
       return residual * ct_b,
@@ -431,7 +435,6 @@ class Jax2TfTest(tf_test_util.JaxToTfTestCase):
     arg = np.arange(3.)
     self.TransformConvertAndCompare(f, arg, "grad")
     # TODO: check that the TF code also computes "sin" 5 times
-
 
   def test_remat_free_var(self):
     def f(x):
@@ -523,7 +526,10 @@ class Jax2TfTest(tf_test_util.JaxToTfTestCase):
       log.append(y.name)
       return x * x
 
-    jax2tf.convert(my_test_function)(2)
+    def caller(x):
+      return my_test_function(jnp.sin(x))
+
+    jax2tf.convert(caller)(2.)
     self.assertIn("my_test_function/foo", log[0])
 
   def test_bfloat16_constant(self):
@@ -593,6 +599,185 @@ class Jax2TfTest(tf_test_util.JaxToTfTestCase):
 
       self.ConvertAndCompare(randint)
 
+  def test_op_metadata_simple(self):
+    self.skipTest("include_xla_op_metadata not yet enabled")
+    # A simple example
+    # The user_frame is used to compute line numbers for ops in the test.
+    user_frame = source_info_util.user_frame(source_info_util.current())
+    def f_simple(x):
+      return jnp.sin(x)
+
+    x = np.ones((2, 3), np.float32)
+    self.CheckOpMetadata(
+        f_simple, x,
+        [tf_test_util.OpMetadataGraph(tf_type="Sin",
+                                      source_file=__file__,
+                                      source_line=user_frame.line_num + 2,
+                                      op_name="jax2tf(f_simple)/sin",
+                                      op_type="sin")
+         ]
+    )
+
+  def test_op_metadata_sub_jit(self):
+    self.skipTest("include_xla_op_metadata not yet enabled")
+    # Calling a jitted-function
+    # The user_frame is used to compute line numbers for ops in the test.
+    user_frame = source_info_util.user_frame(source_info_util.current())
+    def f_callee(x):
+      return jnp.cos(x)
+    def f_caller(x):
+      y = jnp.tanh(x)
+      z = jax.jit(f_callee)(y)
+      return jnp.sin(z)
+
+    x = np.ones((2, 3), np.float32)
+
+    self.CheckOpMetadata(
+        f_caller, x,
+        [tf_test_util.OpMetadataGraph(tf_type="Tanh",
+                                      source_file=__file__,
+                                      source_line=user_frame.line_num + 4,
+                                      op_name="jax2tf(f_caller)/tanh",
+                                      op_type="tanh"),
+         tf_test_util.OpMetadataGraph(tf_type="Cos",
+                                      source_file=__file__,
+                                      source_line=user_frame.line_num + 2,
+                                      op_name="jax2tf(f_caller)/jit(f_callee)/cos",
+                                      op_type="cos"),
+         tf_test_util.OpMetadataGraph(tf_type="Sin",
+                                      source_file=__file__,
+                                      source_line=user_frame.line_num + 6,
+                                      op_name="jax2tf(f_caller)/sin",
+                                      op_type="sin"),
+         ]
+    )
+
+  def test_op_metadata_named(self):
+    self.skipTest("include_xla_op_metadata not yet enabled")
+    # Calling a jax.named_call
+    # The user_frame is used to compute line numbers for ops in the test.
+    user_frame = source_info_util.user_frame(source_info_util.current())
+    def f_callee(x):
+      return jnp.cos(x)
+    def f_caller(x):
+      y = jnp.tanh(x)
+      z = jax.named_call(f_callee, name="callee")(y)
+      return jnp.sin(z)
+
+    x = np.ones((2, 3), np.float32)
+
+    self.CheckOpMetadata(
+        f_caller, x,
+        [tf_test_util.OpMetadataGraph(tf_type="Tanh",
+                                      source_file=__file__,
+                                      source_line=user_frame.line_num + 4,
+                                      op_name="jax2tf(f_caller)/tanh",
+                                      op_type="tanh"),
+         tf_test_util.OpMetadataGraph(tf_type="Cos",
+                                      source_file=__file__,
+                                      source_line=user_frame.line_num + 2,
+                                      op_name="jax2tf(f_caller)/named(callee)/cos",
+                                      op_type="cos"),
+         tf_test_util.OpMetadataGraph(tf_type="Sin",
+                                      source_file=__file__,
+                                      source_line=user_frame.line_num + 6,
+                                      op_name="jax2tf(f_caller)/sin",
+                                      op_type="sin"),
+         ]
+    )
+
+  def test_op_metadata_while_and_cond(self):
+    self.skipTest("include_xla_op_metadata not yet enabled")
+    # An example with while and cond
+    # The user_frame is used to compute line numbers for ops in the test.
+    user_frame = source_info_util.user_frame(source_info_util.current())
+    def f_while_cond(x):
+      def body_fun(i_acc):
+        i, acc = i_acc
+        return (i + 1,
+                (jnp.cos(acc) +
+                 lax.cond(jnp.mod(i, 2) == 0,
+                          lambda acc: jnp.sin(acc),
+                          lambda acc: acc,
+                          acc)))
+
+      _, acc = lax.while_loop(
+          lambda i_acc: i_acc[0] <= 5,
+          body_fun, (0, x))
+      return acc
+
+    x = np.ones((2, 3), np.float32)
+    self.CheckOpMetadata(
+        f_while_cond, x,
+        [tf_test_util.OpMetadataGraph(tf_type="Cos",
+                                      source_file=__file__,
+                                      source_line=user_frame.line_num + 5,
+                                      op_name="jax2tf(f_while_cond)/while/body/cos",
+                                      op_type="cos"),
+         tf_test_util.OpMetadataGraph(tf_type="Sin",
+                                      source_file=__file__,
+                                      source_line=user_frame.line_num + 7,
+                                      op_name="jax2tf(f_while_cond)/while/body/branch_1_fun/sin",
+                                      op_type="sin"),
+         tf_test_util.OpMetadataGraph(tf_type="FloorMod",
+                                      source_file=__file__,
+                                      source_line=user_frame.line_num + 6,
+                                      op_name="jax2tf(f_while_cond)/while/body/rem",
+                                      op_type="rem"),
+         ]
+    )
+
+  def test_op_metadata_batched_while(self):
+    self.skipTest("include_xla_op_metadata not yet enabled")
+    # An example with while and cond
+    # The user_frame is used to compute line numbers for ops in the test.
+    user_frame = source_info_util.user_frame(source_info_util.current())
+    @jax.vmap
+    def f_while(x):
+      def body_fun(carry):
+        new_carry = jnp.sin(carry)  # We look for "sin" in the graph
+        return new_carry
+
+      _, carry = lax.while_loop(
+          lambda carry: jnp.all(carry <= x),  # We look for "le" in the graph
+          body_fun, x)
+      return carry
+
+    shape = (3, 2)
+    x = np.arange(np.prod(shape), dtype=np.float32).reshape(shape)
+
+    jax_comp = jax.xla_computation(f_while)(x)
+    backend = jax.lib.xla_bridge.get_backend()
+    modules = backend.compile(jax_comp).hlo_modules()
+    jax_opt_hlo = modules[0].to_string()
+    print(f"JAX OPT HLO = {jax_opt_hlo}")
+
+    self.CheckOpMetadata(
+        f_while, x,
+        [tf_test_util.OpMetadataGraph(tf_type="Sin",
+                                      source_file=__file__,
+                                      source_line=user_frame.line_num + 4,
+                                      op_name="jax2tf(f_while)/while/body/sin",
+                                      op_type="sin"),
+         tf_test_util.OpMetadataGraph(tf_type="LessEqual",
+                                      source_file=__file__,
+                                      source_line=user_frame.line_num + 8,
+                                      op_name="jax2tf(f_while)/while/body_pred/le",
+                                      op_type="le"),
+         ]
+    )
+
+  def test_op_metadata_disabled(self):
+    self.skipTest("include_xla_op_metadata not yet enabled")
+    def f_simple(x):
+      return jnp.sin(x)
+
+    x = np.ones((2, 3), np.float32)
+    self.CheckOpMetadata(
+        f_simple, x,
+        [],
+        include_xla_op_metadata=False
+    )
 
 if __name__ == "__main__":
   absltest.main(testLoader=jtu.JaxTestLoader())